Refactor join code to reduce duplicated code (#1839)
* Refactor join code to reduce duplicated code


* Move nodeName override to base class
